Remove deadlinks from spanish books (#9671)
* Remove broken link from problem sets

Remove the Caribbean Online Judge site which is not accessible
anymore.

* Remove Spanish book Teoria sintactico-gramatical de objetos

The site gives a 404, and suggests other books (but they are not related to
the topic from the original).

* Remove Spanish book Jardin de JavaScript

The site has disappeared. The author removed the book with the
following commit:
https://github.com/BonsaiDen/JavaScript-Garden/commit/93278f. It seems
like it was unexpected according to its repository collaborators
(https://github.com/BonsaiDen/JavaScript-Garden/issues/403#issuecomment-1682114682).